requirements.txt
openai>=1.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.42.0
docx2pdf>=0.1.8